Prompt_Engineering/requirements.txt
openai>=1.3.0
flask>=2.3.0
python-dotenv>=1.0.0
